LayoutTests/editing/style/font-face-unquote.html

<!DOCTYPE html>
<html>
<body>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script>
description('Test to make sure WebKit doesn\'t quote face attribute of font elements.');

var testContainer = document.createElement('div');
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = 'Hello WebKit';
window.getSelection().selectAllChildren(testContainer);
document.execCommand('fontName', false, 'courier new');

var expected = '<font face="courier new">Hello WebKit</font>';
if (expected == testContainer.innerHTML)
    testPassed('Font face is unquoted.');
else
    testFailed('Font face is quoted, expected: ' + expected + ', actual: ' + testContainer.innerHTML);

document.body.removeChild(testContainer);
var successfullyParsed = true;
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
