LayoutTests/editing/style/font-face-unquote-expected.txt

Test to make sure CyberKit doesn't quote face attribute of font elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Font face is unquoted.
PASS successfullyParsed is true

TEST COMPLETE
